Some cleanup and moving things around
