+++ .eslintrc.json
@@ -1,8 +1,9 @@
 {
-    "extends": "airbnb-base",
-    "rules": {
-      "no-console": "off",
-      "no-await-in-loop": "off",
-      "quote-props": "off"
-    }
+  "extends": "airbnb-base",
+  "rules": {
+    "no-console": "off",
+    "no-await-in-loop": "off",
+    "quote-props": "off",
+    "no-restricted-syntax": "off"
+  }
 }
