+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node index.js",
+    "start": "nodemon src/index.js",
     "lint": "eslint **/*.js"
   },
   "repository": {
